+++ src/main/java/io/anserini/ltr/feature/base/QueryLength.java
@@ -31,7 +31,7 @@ import java.util.List;
 public class QueryLength<T> implements FeatureExtractor<T> {
 
   @Override
-  public Integer extract(Document doc, Terms terms, RerankerContext<T> context) {
+  public float extract(Document doc, Terms terms, RerankerContext<T> context) {
     List<String> queryTokens = context.getQueryTokens();
     return queryTokens.size();
   }
